requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
install==1.3.5
requests==2.28.2
soupsieve==2.4.1
urllib3==1.26.15
python-ebooklib==0.18